refactor(ui): Extract inline sub-components and polish landing/button styles
Changes:
- Extract LogoRow, ReceiptRow and SelectField into their own files (one component per file per AGENTS.md); SelectField is renamed DigitalCardSelectField
- Made PriceHistorySection's content hidden on small screens for responsive design
- Adjusted spacing and padding in StoresMarquee/LogoRow for better alignment and visual consistency
- Updated the free-plan label in landing data for clarity ("Instalacija kao aplikacija")
- Gave the outline button a solid white background for consistent contrast

Splits the three multi-component files into one component per file, plus small responsive and style polish across the landing page.

Notes:
- render-og.tsx and the multi-select compound component were intentionally left as-is (OG image template fragments; shadcn-style compound with shared context).
