+++ frontend/src/app/(root)/components/sections/pricing-receipt.tsx
@@ -1,6 +1,7 @@
 import { ReactNode } from "react";
 import { cn } from "@/lib/utils";
 import EdgeFade from "@/components/custom/common/edge-fade";
+import ReceiptRow from "@/app/(root)/components/sections/receipt-row";
 import type { IPricingRow } from "@/app/(root)/data/landing";
 
 interface IPricingReceiptProps {
@@ -16,21 +17,6 @@ interface IPricingReceiptProps {
 // Free and premium render identically, differing only in copy, badge and dimming.
 const receiptClip =
   "[clip-path:polygon(0_0,100%_0,100%_calc(100%-8px),97%_100%,94%_calc(100%-8px),91%_100%,88%_calc(100%-8px),85%_100%,82%_calc(100%-8px),79%_100%,76%_calc(100%-8px),73%_100%,70%_calc(100%-8px),67%_100%,64%_calc(100%-8px),61%_100%,58%_calc(100%-8px),55%_100%,52%_calc(100%-8px),49%_100%,46%_calc(100%-8px),43%_100%,40%_calc(100%-8px),37%_100%,34%_calc(100%-8px),31%_100%,28%_calc(100%-8px),25%_100%,22%_calc(100%-8px),19%_100%,16%_calc(100%-8px),13%_100%,10%_calc(100%-8px),7%_100%,4%_calc(100%-8px),1%_100%,0_calc(100%-8px))]";
-
-interface IReceiptRowProps {
-  label: string;
-  price: string;
-}
-
-function ReceiptRow({ label, price }: IReceiptRowProps) {
-  return (
-    <li className="flex items-baseline gap-2 text-sm">
-      <span className="text-pretty">{label}</span>
-      <span className="flex-1 border-b border-dotted border-muted-foreground/40" />
-      <span className="font-mono text-pretty">{price}</span>
-    </li>
-  );
-}
 
 export default function PricingReceipt({
   name,
