+++ frontend/src/app/(root)/components/sections/stats-band.tsx
@@ -13,7 +13,7 @@ export default function StatsBand() {
     <section className="relative">
       <h2 className="sr-only">Disscount u brojkama</h2>
 
-      <div className="relative overflow-hidden rounded-3xl bg-primary text-primary-foreground px-6 py-10 sm:py-12 shadow-lg">
+      <div className="relative overflow-hidden rounded-3xl bg-primary text-primary-foreground px-6 py-4 sm:py-8 shadow-lg">
         <SparkleField
           count={6}
           seed="stats-sparkles"
@@ -26,7 +26,7 @@ export default function StatsBand() {
           className="grid grid-cols-1 sm:grid-cols-3 text-center divide-y-2 divide-dashed divide-primary-foreground/40 sm:divide-y-0 sm:divide-x-2"
         >
           {statItems.map((stat) => (
-            <div key={stat.label} className="space-y-1 py-7 sm:px-6">
+            <div key={stat.label} className="space-y-1 py-4 sm:px-6">
               <div className="text-4xl sm:text-5xl font-saira-stencil-semibold text-pretty">
                 {stat.value}
               </div>
