+++ frontend/src/app/(user)/digital-cards/components/forms/digital-card-select-field.tsx
@@ -1,0 +1,62 @@
+"use client";
+
+import { useFormContext } from "react-hook-form";
+
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import type { DigitalCardRequest } from "@/lib/api/types";
+
+interface IDigitalCardSelectFieldProps {
+  name: "type" | "codeType";
+  label: string;
+  placeholder: string;
+  options: { value: string; label: string }[];
+}
+
+export default function DigitalCardSelectField({
+  name,
+  label,
+  placeholder,
+  options,
+}: IDigitalCardSelectFieldProps) {
+  const form = useFormContext<DigitalCardRequest>();
+
+  return (
+    <FormField
+      control={form.control}
+      name={name}
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>{label}</FormLabel>
+          <Select onValueChange={field.onChange} value={field.value}>
+            <FormControl>
+              <SelectTrigger className="w-full">
+                <SelectValue placeholder={placeholder} />
+              </SelectTrigger>
+            </FormControl>
+            <SelectContent>
+              {options.map((option) => (
+                <SelectItem key={option.value} value={option.value}>
+                  {option.label}
+                </SelectItem>
+              ))}
+            </SelectContent>
+          </Select>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
